#2155 cleanup in API functions
